move from_search to WebappIndexer
